# ddtrace/tracer: add config for http status error codes