Add `DD_HTTP_CLIENT_ERROR_STATUSES` and `DD_HTTP_SERVER_ERROR_STATUSES` environment variables for configuring HTTP error codes. Codes are entered as ranges and are parsed accordingly.

Fixes #865.