@nicoledanuwidjaja 2021-03-18
### Benchmark Results
```
name                                 old time/op  new time/op  delta
ParsingClient-8                       102µs ± 0%     6µs ± 0%   ~     (p=1.000 n=1+1)
ParsingServer-8                       109µs ± 0%     7µs ± 0%   ~     (p=1.000 n=1+1)
ParsingClientAndServer-8              200µs ± 0%    13µs ± 0%   ~     (p=1.000 n=1+1)
ParsingFlipperooClient-8              100µs ± 0%     6µs ± 0%   ~     (p=1.000 n=1+1)
ParsingMergeRangesClient-8            477µs ± 0%     7µs ± 0%   ~     (p=1.000 n=1+1)
ParsingMergeRangesClientAndServer-8   955µs ± 0%    14µs ± 0%   ~     (p=1.000 n=1+1)
ParsingFlipperooAndMergeClient-8      248µs ± 0%     7µs ± 0%   ~     (p=1.000 n=1+1)
ParsingClientAndInvalid-8             110µs ± 0%     7µs ± 0%   ~     (p=1.000 n=1+1)
```
**Conclusion:** Using a struct to store the range is significantly faster. (The hypothesis was that it would be faster, but the results proved that this is a **significant** improvement.

[Here](https://github.com/nicoledanuwidjaja/go-dogs/blob/master/tests/) are the results and [modified code](https://github.com/nicoledanuwidjaja/go-dogs/blob/master/tests/set_test.go) that uses the `codeRange` struct.