@richardstartin 2021-03-19
I recommend using a bitset as we do in Java. This has a few advantages:
1. lookups are constant time and cheap which means that classifying status codes is similarly cheap
2. it is implicitly sorted and idempotent, regardless of insertion order, duplicates in input - no need to do sorts or make copies during building
3. the space required can be computed from the max status code in the user defined range, so no resizes during building
4. All server error codes can be represented in 10 `uint64`s = 80 bytes, regardless of density. The client errors require at most 8 `uint64`. This is the same as the space required to store 20 or 16 `uint32` in a sorted array respectively.