@gbbr 2021-03-19
@nicoledanuwidjaja it looks like your code is benchmarking the time it takes to parse the user input, or am I wrong?

What we are looking to benchmark here is how fast / long it takes to check if a certain status code is found in the given range _after_ it was parsed. Parsing the user input happens once at startup, we don't care much about the performance of that. However, checking if a status code passes or not happens _on every request_, we care quite a lot about that.

Also, please post a link to the code for both the benchmarks.