@nicoledanuwidjaja 2021-03-19
Noted -- sorry for the confusion. Here's the benchmarks for **checking if a status code is an error...**

### New Benchmark Results
```
name       old time/op  new time/op  delta
Find599-8  75.1ns ± 0%   6.3ns ± 0%   ~     (p=1.000 n=1+1)
Find499-8  44.2ns ± 0%   6.3ns ± 0%   ~     (p=1.000 n=1+1)
Find400-8  5.72ns ± 0%  6.30ns ± 0%   ~     (p=1.000 n=1+1)
Find500-8  43.5ns ± 0%   6.3ns ± 0%   ~     (p=1.000 n=1+1)
Find600-8  75.2ns ± 0%   5.7ns ± 0%   ~     (p=1.000 n=1+1)
Find450-8  29.9ns ± 0%   6.3ns ± 0%   ~     (p=1.000 n=1+1)
```
The tests were run on two different implementations, differing with the structure to store the parsed HTTP status codes: **old** uses an integer slice, **new** uses an existing [bitset implementation](https://github.com/willf/bitset). 

[Here](https://github.com/nicoledanuwidjaja/go-dogs/blob/master/tests/searching/set_test.go) is the corresponding benchmark tests and code.
[This](https://gist.github.com/nicoledanuwidjaja/f383eba67c33ac2eaa477d70e9615b9f) is the code separated into two respective files.

Thanks @richardstartin for the recommendation! I did also look into the implementation for the de-facto set implementation for Go (ex: `map[int]bool`) as well as the range struct, though the bitset seemed to be faster than both. If that is good, I can add the corresponding code to `globalconfig`.