@gbbr 2021-03-22
@nicoledanuwidjaja, in the code you linked to in your comments, you are benchmarking the time it takes to parse the user input, not to lookup a code in the already parsed user input.