@richardstartin 2021-03-22
I agree that the parsing time is not what's interesting here, it's more the access time that matters, because that happens once per HTTP span.