@gbbr 2021-03-22
A few notes:
* Pass -delta-test none to benchcmp to get actual percentages
* In the first test, don't ignore the error, and check specifically in the benchmark that no error occurs.
* In the second implementation, why are you using a for loop? isn't the whole point of it to use sort.SearchInts ?