@nicoledanuwidjaja 2021-03-22
Great. Took the feedback and ran these tests:
```
name       old time/op  new time/op  delta
Find599-8  75.7ns ± 1%   6.6ns ±13%  -91.23%
Find499-8  45.4ns ± 2%   6.5ns ± 3%  -85.66%
Find400-8  5.91ns ± 3%  6.59ns ± 7%  +11.44%
Find500-8  45.4ns ± 2%   6.5ns ± 5%  -85.60%
Find600-8  77.9ns ± 2%   6.1ns ±21%  -92.14%
Find450-8  30.2ns ± 0%   6.4ns ± 1%  -78.87%
```
Used `benchstat -delta-test none` to generate delta percentages and ran each test 5 times with `go test -bench=. -count 5`.

[Same code](https://github.com/nicoledanuwidjaja/go-dogs/tree/master/tests/searching/test-2), but broke it down a little bit better for clarity.