@gbbr 2021-03-23
Thanks, this is much better. But please also consider:
* Use `sort.SearchInts` to locate the code and benefit of the binary search, not a `for` loop
* Create multiple benchmarks which work on different sets (not always the same set), to have clearer variety. Try using sub-benchmarks (https://blog.golang.org/subtests) for this.
* Check the error in the benchmark and use `b.Fatal` if there is any error returned. Do not ignore string conversion to int errors. Please do not ignore any errors anywhere ever :)